footballleagueback/requirements.txt
Django==3.1.3
djangorestframework
django-cors-headers
django-enumchoicefield